+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Improved the idiomaticity of the `{Instance,Device}Enabled::new` API
 
 ## [0.16.0+162] - 2020-11-24
 ### Changed